+++ src/main/java/com/example/demo/model/FilmText.java
@@ -17,7 +17,7 @@ public class FilmText implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="film_id", columnDefinition = "smallint(5)")
-	private int filmId;
+	private Long filmId;
 
 	@Lob
 	@Column(name="description", columnDefinition = "text")
@@ -28,11 +28,11 @@ public class FilmText implements Serializable {
 	public FilmText() {
 	}
 
-	public int getFilmId() {
+	public Long getFilmId() {
 		return this.filmId;
 	}
 
-	public void setFilmId(int filmId) {
+	public void setFilmId(Long filmId) {
 		this.filmId = filmId;
 	}
